created new function allOptVarIdxs to deal with optVarIdx operations within allVars since some aspects required block to be expanded, but expandBlock requires most of allVars to be finished first.
rerouted optVarIds for state variables.

created field allVars.PKOptVarIdxReroute
n-length vector where n is the optVarIdxs length without state reroutes.
zeros for optVars that do not need to be rerouted.
a nonzero entry 'c' in row 'k' indicates that the k-th column in PK should be rerouted to the c-th column